review: made AbstractProtocolAdapter.thingErrorResponseFromAdaptable abstract and implemented in DittoProtocolAdapter
